requirements.txt
requests
PyYAML
feedparser
python-dateutil
beautifulsoup4
lxml
anthropic
firecrawl-py
python-dotenv